+++ Changelog.md
@@ -11,6 +11,7 @@
 - Enhance resolving of the ndxInitializer to determine the alternating tests.
 
 ### Fixed
+- Problem with seconds in format 'HH:MM:ss' #2745
 - Remove unused mask - Take2 #2748
 - [Question] How to custom placeholder in "ip" alias? #2749
 - Timezone mask #2730